refactor: 관련된 도메인을 모아 패키지로 분리
