+++ src/main/java/blackjack/domain/deck/Card.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 public class Card {
