+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 public class Dealer extends User {
