+++ src/main/java/blackjack/domain/participant/UserName.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 public record UserName(
         String name
