+++ src/main/java/blackjack/domain/participant/Users.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import java.util.List;
 import java.util.function.Consumer;
