+++ src/test/java/blackjack/domain/CardTest.java
@@ -1,9 +1,10 @@
 package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
 
-import org.junit.jupiter.api.DisplayName;
+import blackjack.domain.deck.Card;
+import blackjack.domain.deck.CardShape;
+import blackjack.domain.deck.CardValue;
 import org.junit.jupiter.api.Test;
 
 public class CardTest {
