+++ src/test/java/blackjack/domain/CardsTest.java
@@ -2,6 +2,10 @@ package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.deck.Card;
+import blackjack.domain.deck.CardShape;
+import blackjack.domain.deck.CardValue;
+import blackjack.domain.deck.Cards;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
