+++ src/test/java/blackjack/domain/DealerTest.java
@@ -1,8 +1,11 @@
 package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.deck.Card;
+import blackjack.domain.deck.CardShape;
+import blackjack.domain.deck.CardValue;
+import blackjack.domain.participant.Dealer;
 import org.junit.jupiter.api.Test;
 
 public class DealerTest {
