+++ src/test/java/blackjack/domain/UserTest.java
@@ -3,6 +3,10 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.deck.Card;
+import blackjack.domain.deck.CardShape;
+import blackjack.domain.deck.CardValue;
+import blackjack.domain.participant.User;
 import org.junit.jupiter.api.Test;
 
 public class UserTest {
