+++ src/test/java/blackjack/domain/UsersTest.java
@@ -2,6 +2,11 @@ package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.deck.Card;
+import blackjack.domain.deck.CardShape;
+import blackjack.domain.deck.CardValue;
+import blackjack.domain.participant.User;
+import blackjack.domain.participant.Users;
 import java.util.List;
 import org.junit.jupiter.api.Test;
